refactor(core): rename trace report adapter
Rename the generic trace report adapter and extract the shared text guard so
the file boundaries reflect their actual responsibilities. Also clarify that
the report expansion option opens wrapper spans rather than every span.
